refactor(workbench): move ParseError compoennt
