+++ app/components/workbench/DatasetComponent.tsx
@@ -21,7 +21,7 @@ import Metadata from './Metadata'
 import MetadataEditor from './MetadataEditor'
 import Readme from '../Readme'
 import ReadmeHistory from './ReadmeHistory'
-import ParseError from '../ParseError'
+import ParseError from './ParseError'
 import StatusDot from '../chrome/StatusDot'
 import Structure from '../Structure'
 import SpinnerWithIcon from '../chrome/SpinnerWithIcon'
